Fixed diminishing navigation URL problems.
As it was befor this change, navigation URLs were built using only the
return data to define the new "since" and "until". This breaks because
the window slowly diminishes in size with every navigation. This change
makes the nav URL window use the original user query to define its
bounds UNLESS the query limit was reached.